Marge all PSI tests into single file, move expected code into separate files
