--- components/parser/ktfile/src/test/kotlin/io/github/composegears/valkyrie/parser/ktfile/common/BaseKtParserTest.kt
@@ -1,17 +1,0 @@
-package io.github.composegears.valkyrie.parser.ktfile.common
-
-import com.intellij.openapi.project.Project
-import com.intellij.testFramework.ProjectExtension
-import org.junit.jupiter.api.extension.RegisterExtension
-
-open class BaseKtParserTest {
-
-    companion object {
-        @JvmField
-        @RegisterExtension
-        val projectExtension = ProjectExtension()
-    }
-
-    protected val project: Project
-        get() = projectExtension.project
-}
