+++ components/parser/ktfile/src/test/kotlin/io/github/composegears/valkyrie/parser/ktfile/expected/EmptyImageVector.kt
@@ -1,0 +1,12 @@
+package io.github.composegears.valkyrie.parser.ktfile.expected
+
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+
+val ExpectedEmptyImageVector = ImageVector.Builder(
+    name = "EmptyImageVector",
+    defaultWidth = 24.dp,
+    defaultHeight = 24.dp,
+    viewportWidth = 18f,
+    viewportHeight = 18f,
+).build()
